IRIS-3963 | move auth-facebook and auth-facebookButton tests to auth package
